Add K8S manifests, utilities and example
Add K8S manifest generation using `kustomize` along with an example that
shows how to use it to run a relay instance with stable identity
(instead of the default behaviour where identity is regenerated on start
every time).

Add `skaffold` to stream-line build and deployment of all files and make
it easier to run the whole thing locally. Update readme with
instructions.